64-bit fix

+++ libs/common/sha1.h
@@ -26,8 +26,8 @@
 typedef unsigned char SHA1_DIGEST[SHA1_SIZE];
 
 typedef struct {
-	unsigned long state[5];
-	unsigned long count[2];
+	unsigned int state[5];
+	unsigned int count[2];
 	unsigned char buffer[64];
 } SHA1_CTX;
